Conflicts:
	README.md
	dprr/settings/__init__.py
	templates/includes/nav.html

+++ promrep/signals.py
@@ -1,5 +1,6 @@
-from promrep.models import Group, PostAssertion
+from promrep.models import PostAssertion
 from django.db.models.signals import post_delete
+
 
 def delete_parent_group(sender, instance, *args, **kwargs):
     """
@@ -7,7 +8,8 @@ def delete_parent_group(sender, instance, *args, **kwargs):
     that post only has that post assertion
     """
 
-    if instance.group.postassertion_set.count() == 0:
-        instance.group.delete()
+    if instance.group:
+        if instance.group.postassertion_set.count() == 0:
+            instance.group.delete()
 
 post_delete.connect(delete_parent_group, sender=PostAssertion)
